Move DEFAULT_GRID_SIZE to core/grid
