+++ client/src/core/grid.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_GRID_SIZE = 50;
