+++ client/src/game/dropAsset.ts
@@ -2,6 +2,7 @@ import { assetSystem } from "../assetManager";
 import { assetState } from "../assetManager/state";
 import { clampGridLine, l2gx, l2gy, l2gz } from "../core/conversions";
 import { type GlobalPoint, toGP, Vector } from "../core/geometry";
+import { DEFAULT_GRID_SIZE } from "../core/grid";
 import { baseAdjust } from "../core/http";
 import { SyncMode, InvalidationMode } from "../core/models/types";
 import { uuidv4 } from "../core/utils";
@@ -19,7 +20,6 @@ import type { CharacterId } from "./systems/characters/models";
 import { characterState } from "./systems/characters/state";
 import { floorState } from "./systems/floors/state";
 import { noteSystem } from "./systems/notes";
-import { DEFAULT_GRID_SIZE } from "./systems/position/state";
 import { locationSettingsState } from "./systems/settings/location/state";
 import { selectionBoxFunction } from "./temp";
 import { handleDropFF } from "./ui/firefox";
