+++ client/src/game/shapes/variants/circle.ts
@@ -2,11 +2,10 @@ import type { ApiCircleShape } from "../../../apiTypes";
 import { g2lz, clampGridLine } from "../../../core/conversions";
 import { addP, subtractP, toArrayP, toGP, Vector } from "../../../core/geometry";
 import type { GlobalPoint } from "../../../core/geometry";
+import { DEFAULT_GRID_SIZE } from "../../../core/grid";
 import { FOG_COLOUR } from "../../colour";
-import { calculateDelta } from "../../drag";
 import type { GlobalId, LocalId } from "../../id";
 import type { IShape } from "../../interfaces/shape";
-import { DEFAULT_GRID_SIZE } from "../../systems/position/state";
 import { getProperties } from "../../systems/properties/state";
 import type { ShapeProperties } from "../../systems/properties/state";
 import { Shape } from "../shape";
