+++ client/src/game/systems/position/state.ts
@@ -2,8 +2,6 @@ import type { LocalPoint } from "../../../core/geometry";
 import type { LocalId } from "../../id";
 import { buildState } from "../state";
 
-export const DEFAULT_GRID_SIZE = 50;
-
 const state = buildState(
     {
         outOfBounds: false,
