+++ client/src/game/tools/variants/map.ts
@@ -3,6 +3,7 @@ import { reactive } from "vue";
 import { l2g } from "../../../core/conversions";
 import { addP, cloneP, subtractP, toGP, Vector } from "../../../core/geometry";
 import type { GlobalPoint, LocalPoint } from "../../../core/geometry";
+import { DEFAULT_GRID_SIZE } from "../../../core/grid";
 import { InvalidationMode, SyncMode } from "../../../core/models/types";
 import { i18n } from "../../../i18n";
 import { sendShapePositionUpdate, sendShapeSizeUpdate } from "../../api/emits/shape/core";
@@ -12,7 +13,6 @@ import { ToolName } from "../../models/tools";
 import type { ITool, ToolPermission } from "../../models/tools";
 import { Rect } from "../../shapes/variants/rect";
 import { floorState } from "../../systems/floors/state";
-import { DEFAULT_GRID_SIZE } from "../../systems/position/state";
 import { selectedSystem } from "../../systems/selected";
 import { selectedState } from "../../systems/selected/state";
 import { SelectFeatures } from "../models/select";
